new try

+++ hashmap.c
@@ -147,7 +147,7 @@ Pair * firstMap(HashMap * map) {
 Pair * nextMap(HashMap * map) {
 
   long nextToCurrent = map->current + 1;
-  for (int i = nextToCurrent; i < map->capacity; i++){
+  for (long i = nextToCurrent; i < map->capacity; i++){
     if ((map->buckets[i] != NULL) && (map->buckets[i]->key != NULL)){
       map->current = i;
       return map->buckets[i];
